Convert municipality number to Int. Change MunicipalityProvider to accept set of ELY numbers.

+++ digiroad2-geo/src/main/scala/fi/liikennevirasto/digiroad2/user/User.scala
@@ -4,7 +4,7 @@ case class Configuration(
     zoom: Option[Long] = None,
     east: Option[Long] = None,
     north: Option[Long] = None,
-    municipalityNumber: Option[Long]  = None,
-    authorizedMunicipalities: Set[Long] = Set()
+    municipalityNumber: Option[Int]  = None,
+    authorizedMunicipalities: Set[Int] = Set()
 )
 case class User(id: Long, username: String, configuration: Configuration)
